feat: add design and assets to the project
Added in the design that will be built up in this project.

Import fonts from google fonts.

Added in image assets that are going to be of use.

Added the favicon.

Mixins are added defining the breakpoints to be used in this project.

Colors are added as scss variables.

Added in a template readme file which will be used  to explain the project.